+++ Common/Database/Dbc/EmotesTextReader.cs
@@ -1,0 +1,12 @@
+﻿using System.Linq;
+
+namespace Common.Database.Dbc
+{
+    public class EmotesTextReader : DbcReader<EmotesText>
+    {
+        public EmotesText GetData(int id)
+        {
+            return RecordDataIndexed.Values.ToArray().FirstOrDefault(a => a.Id == id);
+        }
+    }
+}