+++ Common/Database/Dbc/MapReader.cs
@@ -1,0 +1,12 @@
+﻿using System.Linq;
+
+namespace Common.Database.Dbc
+{
+    public class MapReader : DbcReader<Map>
+    {
+        public Map GetArea(int id)
+        {
+            return RecordDataIndexed.Values.ToArray().FirstOrDefault(a => a.MapId == id);
+        }
+    }
+}