+++ Common/Database/Xml/RacesXml.cs
@@ -2,361 +2,361 @@
 
 namespace Common.Database.Xml
 {
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     [XmlRoot(IsNullable = false, ElementName = "races")]
     public class RacesXml
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlElement("race")]
         public racesRace[] race { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRace
     {
-        /// <remarks/>
+        /// <remarks />
         public byte health { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceInit init { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceStats stats { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceLeveling leveling { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("spell", IsNullable = false)]
         public racesRaceSpell[] spells { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("skill", IsNullable = false)]
         public racesRaceSkill[] skills { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("action", IsNullable = false)]
         public racesRaceAction[] actions { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("class", IsNullable = false)]
         public racesRaceClass[] classes { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte id { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceInit
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte MapId { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte ZoneId { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public float MapX { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public float MapY { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public float MapZ { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public float MapR { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte Cinematic { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceStats
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte str { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte agi { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte sta { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte @int { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte spi { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceLeveling
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal str { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal agi { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal sta { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal @int { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal spi { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceSpell
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort id { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceSkill
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort id { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort min { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort max { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceAction
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte button { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort action { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte type { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClass
     {
-        /// <remarks/>
+        /// <remarks />
         public byte health { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public string atackspeed { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassStats stats { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassLeveling leveling { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassModificadores modificadores { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassPower power { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("spell", IsNullable = false)]
         public racesRaceClassSpell[] spells { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("skill", IsNullable = false)]
         public racesRaceClassSkill[] skills { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlArrayItem("action", IsNullable = false)]
         public racesRaceClassAction[] actions { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public string id { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassStats
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte str { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte agi { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte sta { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte @int { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte spi { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassLeveling
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal str { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal agi { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal sta { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal @int { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public decimal spi { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassModificadores
     {
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassModificadoresSta sta { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         public racesRaceClassModificadoresInt @int { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassModificadoresSta
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte @base { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte mod { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassModificadoresInt
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte @base { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte mod { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassPower
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte mana { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort rage { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte energy { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassSpell
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort id { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassSkill
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort id { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte min { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte max { get; set; }
     }
 
-    /// <remarks/>
+    /// <remarks />
     [XmlType(AnonymousType = true)]
     public class racesRaceClassAction
     {
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte button { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public ushort action { get; set; }
 
-        /// <remarks/>
+        /// <remarks />
         [XmlAttribute]
         public byte type { get; set; }
     }