+++ RealmServer/Enums/InventorySlots.cs
@@ -1,0 +1,49 @@
+﻿namespace RealmServer.Enums
+{
+    public enum InventorySlots
+    {
+        SLOT_START = 0,
+        SLOT_HEAD = 0,
+        SLOT_NECK = 1,
+        SLOT_SHOULDERS = 2,
+        SLOT_SHIRT = 3,
+        SLOT_CHEST = 4,
+        SLOT_WAIST = 5,
+        SLOT_LEGS = 6,
+        SLOT_FEET = 7,
+        SLOT_WRISTS = 8,
+        SLOT_HANDS = 9,
+        SLOT_FINGERL = 10,
+        SLOT_FINGERR = 11,
+        SLOT_TRINKETL = 12,
+        SLOT_TRINKETR = 13,
+        SLOT_BACK = 14,
+        SLOT_MAINHAND = 15,
+        SLOT_OFFHAND = 16,
+        SLOT_RANGED = 17,
+        SLOT_TABARD = 18,
+        SLOT_END = 19,
+
+        // Misc Types
+        SLOT_BAG_START = 19,
+        SLOT_BAG1 = 19,
+        SLOT_BAG2 = 20,
+        SLOT_BAG3 = 21,
+        SLOT_BAG4 = 22,
+        SLOT_INBACKPACK = 23,
+        SLOT_BAG_END = 23,
+
+        SLOT_ITEM_START = 23,
+        SLOT_ITEM_END = 39,
+
+        SLOT_BANK_ITEM_START = 39,
+        SLOT_BANK_ITEM_END = 63,
+        SLOT_BANK_BAG_1 = 63,
+        SLOT_BANK_BAG_2 = 64,
+        SLOT_BANK_BAG_3 = 65,
+        SLOT_BANK_BAG_4 = 66,
+        SLOT_BANK_BAG_5 = 67,
+        SLOT_BANK_BAG_6 = 68,
+        SLOT_BANK_END = 69
+    }
+}