+++ RealmServer/Handlers/OnCharCreate.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 using Common.Globals;
 using Common.Helpers;
 using RealmServer.Database;
@@ -9,28 +10,29 @@ namespace RealmServer.Handlers
 {
     public class OnCharCreate
     {
-        internal static void Handler(RealmServerSession session, CMSG_CHAR_CREATE handler)
+        internal static async void Handler(RealmServerSession session, CMSG_CHAR_CREATE handler)
         {
             session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_SUCCESS));
 
+            // If limit character reached
+            if (Characters.GetCharacters(session.User).Count >= Config.Instance.LimitCharacterRealm)
+            {
+                session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_SERVER_LIMIT));
+                return;
+            }
+
             try
             {
-                // If limit character reached
-                if (Characters.GetCharacters(session.User).Count >= Config.Instance.LimitCharacterRealm)
-                {
-                    session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_SERVER_LIMIT));
-                    return;
-                }
-
-                // check if name in use
-
                 Characters.Create(handler, session.User);
-
-                session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_SUCCESS));
             }
             catch (Exception)
             {
                 session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_ERROR));
+            }
+            finally
+            {
+                await Task.Delay(2500);
+                session.SendPacket(new SMSG_CHAR_CREATE(LoginErrorCode.CHAR_CREATE_SUCCESS));
             }
         }
     }