+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.12] - 2018-10-15
+### Added
+- Character Table
+  - Flag = CharacterFlag(Enum)
+  - Cinematic = bool
+- Item Table added.
+- RealmServer
+  - Added InventorySlots Enum.
+
+### Changed
+- Common DBC Splited to Matching Files.
+- RealmServer: SMSG_CHAR_ENUM now show item gear.
+
 ## [0.0.11] - 2018-10-15
 ### Added
 - RealmServer
@@ -38,7 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - CMSG_SET_ACTIVE_MOVER
   - MSG_MOVE_FALL_LAND
 - World Entity
-  
+
 ### Changed
 - RealmServerSession users to user.